Add comments and (some) tests.
