+++ clinvar/src/main/scala/org/broadinstitute/monster/etl/clinvar/ClinVarSplitters.scala
@@ -1,0 +1,257 @@
+package org.broadinstitute.monster.etl.clinvar
+
+import java.util.concurrent.atomic.AtomicInteger
+
+import com.spotify.scio.coders.Coder
+import com.spotify.scio.values.{SCollection, SideOutput}
+import upack._
+
+import scala.collection.mutable
+
+/** Container for functions used to split apart ClinVar's nested data. */
+object ClinVarSplitters {
+  import ClinVarContants._
+
+  def getVariant(message: Msg): Option[Msg] =
+    VariantTypes.foldLeft(Option.empty[Msg]) { (acc, subtype) =>
+      acc.orElse {
+        message.obj.remove(subtype).map {
+          case arr @ Arr(msgs) =>
+            msgs.foreach(_.obj.update(SubclassKey, subtype))
+            arr
+          case msg =>
+            msg.obj.update(SubclassKey, subtype)
+            msg
+        }
+      }
+    }
+
+  def unrollVariants(
+    variantWrapper: Msg,
+    parentIds: List[Msg],
+    getId: Msg => Msg,
+    output: Msg => Unit
+  ): Iterable[Msg] =
+    getVariant(variantWrapper).toIterable.flatMap {
+      case Arr(vs) => vs
+      case other   => Iterable(other)
+    }.map { variant =>
+      val variantId = getId(variant)
+      val immediateParentId = parentIds.headOption
+
+      // Unroll any children.
+      val _ = unrollVariants(variant, variantId :: parentIds, getId, output)
+
+      // Link upwards.
+      variant.obj.update(ParentsRef, Arr(parentIds.to[mutable.ArrayBuffer]))
+      immediateParentId.foreach(variant.obj.update(ParentRef, _))
+
+      // Output this fully-unrolled variant.
+      output(variant)
+
+      // Return this ID, in case the caller cares.
+      variantId
+    }
+
+  def splitArchives(fullVcvStream: SCollection[Msg])(implicit coder: Coder[Msg]): (
+    SCollection[Msg],
+    SCollection[Msg],
+    SCollection[Msg],
+    SCollection[Msg],
+    SCollection[Msg]
+  ) = {
+    val rcvOut = SideOutput[Msg]
+    val variationOut = SideOutput[Msg]
+    val scvOut = SideOutput[Msg]
+    val scvVariationOut = SideOutput[Msg]
+
+    val (trimmedVcvStream, sideCtx) = fullVcvStream
+      .withSideOutputs(
+        rcvOut,
+        variationOut,
+        scvOut,
+        scvVariationOut
+      )
+      .withName("Split Variation Archives")
+      .map { (fullVcv, ctx) =>
+        // Init top-level VCV info.
+        val vcvObj = fullVcv.obj
+        val vcvId = vcvObj(Str("@Accession"))
+        val trimmedOut = new mutable.LinkedHashMap[Msg, Msg]
+
+        // Pull out top-level record.
+        // Fields which are destined for other tables will be stripped away.
+        // Whatever's left behind will be re-added to the trimmed VCV to
+        // preserve all information.
+        def processRecord(originalRecord: Msg): Msg = {
+          val recordCopy = upack.copy(originalRecord)
+
+          // Extract and push out any RCVs in the VCV, tracking the collected IDs.
+          val rcvIds = {
+            val idBuffer = new mutable.ArrayBuffer[Msg]()
+            val rcvs = for {
+              wrapper <- recordCopy.obj.remove(Str("RCVList"))
+              arrayOrSingle <- wrapper.obj.get(Str("RCVAccession"))
+            } yield {
+              arrayOrSingle match {
+                case Arr(msgs) => msgs
+                case msg       => Iterable(msg)
+              }
+            }
+            rcvs.getOrElse(Iterable.empty).foreach { rcv =>
+              val rcvObj = rcv.obj
+
+              // Generate and track RCV ID.
+              val id = rcvObj(Str("@Accession"))
+              idBuffer.append(id)
+
+              // Link back to VCV.
+              rcvObj.update(VcvRef, vcvId)
+
+              // Push to side output.
+              ctx.output(rcvOut, rcv)
+            }
+
+            Arr(idBuffer)
+          }
+
+          // Extract and push top-level variant info from the VCV.
+          val variationId = unrollVariants(
+            recordCopy,
+            Nil,
+            _.obj(Str("@VariationID")),
+            variant => {
+              // Link back to the VCV and RCVs.
+              variant.obj.update(VcvRef, vcvId)
+              variant.obj.update(RcvRef, rcvIds)
+              // Push to side output.
+              ctx.output(variationOut, variant)
+              ()
+            }
+          )
+
+          // Extract and push out any SCVs and SCV-level variations.
+          val scvs = for {
+            wrapper <- recordCopy.obj.remove(Str("ClinicalAssertionList"))
+            arrayOrSingle <- wrapper.obj.get(Str("ClinicalAssertion"))
+          } yield {
+            arrayOrSingle match {
+              case Arr(msgs) => msgs
+              case msg       => Iterable(msg)
+            }
+          }
+          scvs.getOrElse(Iterable.empty).foreach { scv =>
+            val scvObj = scv.obj
+            val scvId = scvObj(Str("ClinVarAccession")).obj(Str("@Accession"))
+
+            // Link the SCV back to the VCV, RCVs, and top-level variant.
+            scvObj.update(VcvRef, vcvId)
+            scvObj.update(RcvRef, rcvIds)
+            variationId.foreach(scvObj.update(VarRef, _))
+
+            // Extract out any SCV-level variation.
+            val counter = new AtomicInteger
+            val _ = unrollVariants(
+              scv,
+              Nil,
+              variant => {
+                val nextId = Str(s"${scvId.str}.${counter.getAndIncrement()}")
+                variant.obj.update(IdKey, nextId)
+                nextId
+              },
+              variant => {
+                // Link the variation back to its SCV.
+                variant.obj.update(ScvRef, scvId)
+                // Push to side output.
+                ctx.output(scvVariationOut, variant)
+                ()
+              }
+            )
+
+            // Push SCV to side output.
+            ctx.output(scvOut, scv)
+          }
+
+          recordCopy
+        }
+
+        // TODO useful comment
+        vcvObj.foreach {
+          case (k, v) =>
+            if (RecordTypes.contains(k)) {
+              trimmedOut.update(k, processRecord(v))
+            } else {
+              trimmedOut.update(k, v)
+            }
+        }
+
+        // Whatever's left in the VCV object is returned as the main output.
+        Obj(trimmedOut): Msg
+      }
+
+    (
+      trimmedVcvStream,
+      sideCtx(rcvOut),
+      sideCtx(variationOut),
+      sideCtx(scvOut),
+      sideCtx(scvVariationOut)
+    )
+  }
+
+  def splitScvs(scvs: SCollection[Msg])(
+    implicit coder: Coder[Msg]
+  ): (SCollection[Msg], SCollection[Msg], SCollection[Msg]) = {
+    val submitterOut = SideOutput[Msg]
+    val submissionOut = SideOutput[Msg]
+
+    val (main, side) =
+      scvs.withSideOutputs(submitterOut, submissionOut).withName("Split SCVs").map {
+        (scv, ctx) =>
+          val scvObj = scv.obj
+          val newScv = new mutable.LinkedHashMap[Msg, Msg]()
+
+          val scvId = scvObj(Str("accession"))
+          val orgId = scvObj(Str("org_id"))
+          val submitDate = scvObj(Str("submission_date"))
+          val submissionId = Str(s"${orgId.str}.${submitDate.str}")
+
+          val submitter = new mutable.LinkedHashMap[Msg, Msg]()
+          val submission = new mutable.LinkedHashMap[Msg, Msg]()
+
+          // Set and link IDs
+          submitter.update(IdKey, orgId)
+          submission.update(IdKey, submissionId)
+          submission.update(SubmitterRef, orgId)
+          newScv.update(IdKey, scvId)
+          newScv.update(SubmitterRef, orgId)
+          newScv.update(SubmissionRef, submissionId)
+
+          // Fill in other properties
+          val submitterFields =
+            Set("submitter_name", "org_category", "org_abbrev").map(Str)
+          val submissionFields =
+            Set("submission_date", "submission_names").map(Str)
+          (scvObj.keySet -- submitterFields -- submissionFields - Str("org_id")).foreach {
+            k =>
+              scvObj.get(k).foreach(newScv.update(k, _))
+          }
+          submitterFields.foreach { k =>
+            scvObj.get(k).foreach(submitter.update(k, _))
+          }
+          submissionFields.foreach { k =>
+            scvObj.get(k).foreach(submission.update(k, _))
+          }
+
+          // Push outputs.
+          ctx.output(submitterOut, Obj(submitter): Msg)
+          ctx.output(submissionOut, Obj(submission): Msg)
+          Obj(newScv): Msg
+      }
+
+    (
+      main,
+      side(submitterOut).distinctBy(_.obj(IdKey).str),
+      side(submissionOut).distinctBy(_.obj(IdKey).str)
+    )
+  }
+}
